# api/openapi.yaml
openapi: 3.1.0
info:
  title: Zoomies API
  version: "1.0.0"
  summary: Self-hosted GitHub Actions runner fleet controller
  description: |
    Zoomies manages ephemeral self-hosted GitHub Actions runners.

    This document is the whole product surface. The web UI and the `zoomies`
    CLI are both clients of it, and nothing is reachable from either that is
    not described here.

    ## Authentication

    Two credentials are accepted on the same routes:

    * a **session cookie** (`zoomies_session`), which the browser gets from
      `POST /auth/login`;
    * a **bearer token** (`Authorization: Bearer zoo_...`), for automation.

    Cookie-authenticated mutating requests are subject to a same-origin check.
    Bearer-authenticated requests are not, because they are not vulnerable to
    CSRF.

    Agents use a separate credential class on `/agent/*` and can reach nothing
    else.

    ## Roles

    | Role | May |
    |---|---|
    | `viewer` | read everything except secrets |
    | `operator` | act on the fleet and manage pools |
    | `admin` | manage users, tokens, installations and settings |

    Each operation below records its minimum role in `x-zoomies-role`.

    ## Conventions

    * List endpoints take `limit`, `offset`, `sort` and `order`, and return
      `{ items, total, limit, offset }`.
    * Timestamps are RFC 3339 with a `Z` offset.
    * Durations are Go duration strings: `"5m"`, `"1h30s"`.
...
      message written for a human to read.
  license:
...
servers:
...
security:
  - cookieAuth: []
  - bearerAuth: []
...
paths:
...
  /auth/bootstrap:
    post:
      operationId: bootstrap
      tags: [auth]
      summary: Create the first administrator
      description: |
        Available only while no user exists. Once any account has been created
        this returns 409 forever. That check is the entire security of this
        route, which is why it is unauthenticated.
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [username, password]
              properties:
                username: { type: string, minLength: 1, maxLength: 64, examples: [admin] }
                password: { type: string, minLength: 12, format: password }
                email: { type: string, format: email }
      responses:
        "201":
          description: Created, and the session cookie is set
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Identity" }
        "409": { $ref: "#/components/responses/Conflict" }
        "422": { $ref: "#/components/responses/Unprocessable" }
...
  /users/{id}:
    parameters: [{ $ref: "#/components/parameters/PathID" }]
    get:
      operationId: getUser
      tags: [admin]
      summary: Get a user
      x-zoomies-role: admin
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema: { $ref: "#/components/schemas/User" }
        "404": { $ref: "#/components/responses/NotFound" }
    patch:
      operationId: updateUser
      tags: [admin]
      summary: Update a user
      description: Refused if it would leave the instance with no enabled administrator.
      x-zoomies-role: admin
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                email: { type: string, format: email }
                display_name: { type: string }
                role: { $ref: "#/components/schemas/Role" }
                disabled: { type: boolean }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema: { $ref: "#/components/schemas/User" }
        "409": { $ref: "#/components/responses/Conflict" }
    delete:
      operationId: deleteUser
      tags: [admin]
      summary: Delete a user
      description: Refused if it would leave the instance with no enabled administrator.
      x-zoomies-role: admin
      responses:
        "204": { description: Deleted }
        "409": { $ref: "#/components/responses/Conflict" }
...
components:
  securitySchemes:
    cookieAuth:
      type: apiKey
      in: cookie
      name: zoomies_session
      description: Set by `POST /auth/login`. HttpOnly, SameSite=Lax, and Secure when the external URL is https.
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: zoo_<prefix>_<secret>
      description: An API token. Not subject to the same-origin check, because it is not vulnerable to CSRF.
  parameters:
    PathID:
      name: id
      in: path
      required: true
      schema: { type: string }
      description: The resource ID, e.g. `pool_k3f9qz2m`.
...
  responses:
...
    NotFound:
      description: No such thing.
      content:
        application/json:
          schema: { $ref: "#/components/schemas/ErrorEnvelope" }
    Conflict:
      description: The request conflicts with the current state.
      content:
        application/json:
          schema: { $ref: "#/components/schemas/ErrorEnvelope" }
    Unprocessable:
      description: The request was understood but is not valid. `errors` names the fields.
      content:
        application/json:
          schema:
            allOf:
              - $ref: "#/components/schemas/ErrorEnvelope"
              - type: object
                properties:
                  errors: { type: array, items: { $ref: "#/components/schemas/FieldError" } }
...
  schemas:
...
    Role:
      type: string
      enum: [viewer, operator, admin]
...
    ErrorEnvelope:
      type: object
      required: [error]
      properties:
        error:
          type: object
          required: [code, message]
          properties:
            code:
              type: string
              enum: [bad_request, unauthorized, forbidden, not_found, conflict, unprocessable, rate_limited, internal]
            message: { type: string, description: Written for a person to read, and to act on. }
            field: { type: string }
            detail: { type: string }
    FieldError:
      type: object
      required: [field, message]
      properties:
        field: { type: string, examples: [labels] }
        message: { type: string, examples: ["a pool needs at least one label your workflows can ask for"] }
...
    User:
      type: object
      properties:
        id: { type: string }
        username: { type: string }
        email: { type: string }
        display_name: { type: string }
        role: { $ref: "#/components/schemas/Role" }
        oidc_subject: { type: string }
        disabled: { type: boolean }
        must_change_password: { type: boolean }
        created_at: { type: string, format: date-time }
        last_login_at: { type: [string, "null"], format: date-time }
    Identity:
      type: object
      properties:
        kind: { type: string, enum: [user, token, agent, system] }
        id: { type: string }
        name: { type: string }
        role: { $ref: "#/components/schemas/Role" }
        scopes: { type: array, items: { type: string } }
        must_change_password: { type: boolean }
...